Enable 'Preserve Mesh Hierarchy' option by default
The Jones3D engine has a hardcoded bug where, when attaching one thing to another thing's mesh, it uses the mesh number as the joint number in the hierarchy node. Coincidentally, the mesh number and its hierarchy node number match in the original 3DO files. To avoid issues caused by this bug, the 'Preserve Mesh Hierarchy' option is enabled by default with this commit.